+++ packages/expo-notifications/src/setNotificationCategoryAsync.ts
@@ -14,16 +14,17 @@ import {
  * > Don't use the characters `:` or `-` in your category identifier. If you do, categories might not work as expected.
  * @param actions An array of [`NotificationAction`s](#notificationaction), which describe the actions associated with this category.
  * @param options An optional object of additional configuration options for your category.
- * @return A Promise which resolves to the category you just have created.
+ * @return A Promise which resolves to the category you just have created, or null on web
  * @platform android
  * @platform ios
+ * @platform web
  * @header categories
  */
 export default async function setNotificationCategoryAsync(
   identifier: string,
   actions: NotificationAction[],
   options?: NotificationCategoryOptions
-): Promise<NotificationCategory | null> {
+): Promise<NotificationCategory> {
   if (!NotificationCategoriesModule.setNotificationCategoryAsync) {
     throw new UnavailabilityError('Notifications', 'setNotificationCategoryAsync');
   }