+++ packages/expo-notifications/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### 🐛 Bug fixes
 
+- Throw `UnavailabilityError` when trying to use `setNotificationCategoryAsync` on web.
 - Remove `.native` hardcoded platform imports ([#26511](https://github.com/expo/expo/pull/26511) by [@marklawlor](https://github.com/marklawlor))
 
 ### 💡 Others